[#182] Refacotr: 커스텀 훅 분리
